requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50

# PDF processing
pypdf>=3.17.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# LLM integration (OpenRouter uses OpenAI SDK)
openai>=1.10.0

# Web interface
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# Progress tracking
tqdm>=4.66.0

